refactor(mcp): extract pure row helpers from panel.ts into panel-rows.ts
